bot/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Telegram Bot
python-telegram-bot>=20.7

# AI & HTTP
openai>=1.10.0
aiohttp>=3.9.0

# Database
supabase>=2.3.0

# Utils
python-dotenv>=1.0.0
loguru>=0.7.0
